requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
schedule==1.2.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
json5==0.9.14
pydantic==2.5.2
tenacity==8.2.3
chromadb==0.4.22
openai==1.12.0
tiktoken==0.5.2
fastapi==0.109.0
uvicorn==0.27.0

